### 2024-04-01 #6565 @danielskatz
@JIBSIL - thanks for your submission. I notice that your paper doesn't seem to have any references, and is extremely short. Can you expand it a bit, including some related work perhaps? Please see https://joss.readthedocs.io/en/latest/submitting.html#how-should-my-paper-be-formatted and https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
