### 2024-04-01 95bd602
Commit paper revision + bibliography
