+++ PAPER.md
@@ -13,17 +13,27 @@ affiliations:
  - name: Independent Researcher, United States
    index: 1
 date: 31 March 2024
+bibliography: paper.bib
 ---
 
 # Summary
 
-In an age where the distribution of information is crucial, current file sharing solutions suffer significant deficiencies. Popular systems such as Google Drive, torrenting and IPFS suffer issues with compatibility, accessibility and censorship. DistriFS provides a novel decentralized approach tailored for efficient and large-scale distribution of files. The proposed server implementation harnesses the power of Go, ensuring near-universal interoperability across operating systems and hardware. Moreover, the use of the HTTP protocol eliminates the need for additional software to access the network, ensuring compatibility across all major operating systems and facilitating effortless downloads. The design and efficacy of DistriFS represent a significant advancement in the realm of file distribution systems, offering a scalable and secure alternative to current centralized and decentralized models.
+In an age where the distribution of information is crucial, current file sharing solutions suffer significant deficiencies. Popular systems such as Google Drive, torrenting and IPFS suffer issues with compatibility, accessibility and censorship. DistriFS provides a novel decentralized approach tailored for efficient and large-scale distribution of files. The proposed server implementation harnesses the power of Go, ensuring near-universal interoperability across operating systems and hardware [^1]. Moreover, the use of the HTTP protocol eliminates the need for additional software to access the network, ensuring compatibility across all major operating systems and facilitating effortless downloads. The design and efficacy of DistriFS represent a significant advancement in the realm of file distribution systems, offering a scalable and secure alternative to current centralized and decentralized models.
+
+# Current Work
+
+Many researchers and laypeople alike choose existing decentralized solutions to distribute large files, such as torrenting and the InterPlanetary FileSystem (IPFS). While such alternatives to centralized file-sharing services are in use, their implementation often falls short of the user-friendly experience offered by centralized counterparts. Torrenting is often impeded by firewall restrictions, due to an assumption made by many governments and ISPs that torrenting traffic is solely used for downloading illegal content. The majority of users are deterred from using torrenting due to these barriers, often resorting to paid VPNs or abandoning the method entirely [Morris:2009]. Other more recent solutions, such as IPFS, circumvent firewalls and government censorship more effectively. However, they lack ease-of-use and accessibility, and demonstrate inefficiencies in downloading less popular files [Benet:2014]. Shortcomings in accessibility exclude a minority of users, particularly those with disabilities and those using non-mainstream operating systems and hardware [Burda:2013]. While IPFS employs a similar architecture to DistriFS with the use of browser-based downloads, it introduces additional complexity by needing to translate these HTTP requests into its native TCP protocol. This translation often results in downtime and timeout issues, particularly under heavy traffic conditions [Wan:2017]. In the academic sphere, studies like "Frangipani" [Thekkath:1997] have delved into decentralized file systems, examining their potential and limitations. However, these studies have not fully addressed the specific challenges of creating a practical system that is both user-friendly and privacy-respecting, a key focus of DistriFS.
+
+The most technically inclined of users may lean towards self-hosted platforms such as OwnCloud, NextCloud and Seafile to overcome the limitations and risk of trusting closed-source platforms. However, these alternatives are still centralized, and thus are vulnerable to a different set of data loss issues. Self-hosted platforms are vulnerable to physical drive failure, natural disasters and ransomware. While users are recommended to take mitigation steps like keeping regular backups and monitoring hard-drive health, very few individuals can afford ISO business continuity certifications [^2] and professional audits to verify the security of their systems.
 
 # Statement of Need
 
[…]
