+++ paper.bib
@@ -1,0 +1,69 @@
+@inproceedings{Johnson:2008,
+	title        = {The evolution of the peer-to-peer file sharing industry and the security risks for users},
+	author       = {Johnson, M. E. and McGuire, D. and Willey, N. D.},
+	year         = 2008,
+	booktitle    = {Proceedings of the 41st Annual Hawaii International Conference on System Sciences (HICSS 2008)},
+	address      = {Waikoloa, HI, USA},
+	pages        = 383,
+	doi          = {10.1109/HICSS.2008.436},
+	url          = {https://doi.org/10.1109/HICSS.2008.436}
+}
+@article{Burda:2013,
+	title        = {Sustaining accessibility of information through digital preservation: A literature review},
+	author       = {Burda, D. and Teuteberg, F.},
+	year         = 2013,
+	journal      = {Journal of Information Science},
+	volume       = 39,
+	pages        = {442--458},
+	doi          = {10.1177/0165551513480107},
+	url          = {https://doi.org/10.1177/0165551513480107}
+}
+@article{Morris:2009,
+	title        = {Pirates of the Internet: At intellectual property's end with torrents and challenges for choice of law},
+	author       = {Morris, P. S.},
+	year         = 2009,
+	journal      = {International Journal of Law and Information Technology},
+	volume       = 17,
+	number       = 3,
+	pages        = {282--303},
+	doi          = {10.1093/ijlit/ean010},
+	url          = {https://doi.org/10.1093/ijlit/ean010}
+}
+@misc{Benet:2014,
+	title        = {IPFS - Content addressed, versioned, P2P file system},
+	author       = {Benet, J.},
+	year         = 2014,
+	url          = {https://arxiv.org/abs/1407.3561},
+	eprint       = {1407.3561},
+	archiveprefix = {arXiv},
+	primaryclass = {cs.NI}
+}
+@article{Thekkath:1997,
+	title        = {Frangipani: A scalable distributed file system},
+	author       = {Thekkath, C. A. and Mann, T. and Lee, E. K.},
+	year         = 1997,
+	journal      = {SIGOPS Operating Systems Review},
+	volume       = 31,
+	number       = 5,
+	pages        = {224--237},
+	doi          = {10.1145/269005.266694},
+	url          = {https://doi.org/10.1145/269005.266694}
+}
+@article{Cox:2022,
+	title        = {The Go programming language and environment},
+	author       = {Lu, L. and Sankaranarayana Pillai, T. and Gopalakrishnan, H. and Arpaci-Dusseau, A. C. and Cox, R. and Griesemer, R. and Pike, R. and Taylor, I. L. and Thompson, K.},
+	year         = 2022,
+	journal      = {Communications of the ACM},
+	volume       = 65,
+	number       = 5,
+	pages        = {70--78},
+	doi          = {10.1145/3488716},
+	url          = {https://doi.org/10.1145/3488716}
+}
+@inproceedings{Wan:2017,
+	title        = {Bug characteristics in blockchain systems: A large-scale empirical study},
+	author       = {Wan, Z. and Lo, D. and Xia, X. and Cai, L.},
+	year         = 2017,
+	booktitle    = {Proceedings of the 14th IEEE International Working Conference on Mining Software Repositories: MSR},
+	pages        = {20--21}
+}
