### 2024-04-01 #6565 @JIBSIL
@danielskatz - I've added a Current Work section, bibliography and some references from the original paper. Apologies for the lack of references in the original - this is my first time converting a paper into a shorter form. Thanks for checking it out!
