### 2024-04-01 #6565 @xuanxu
editorilabot is case sensitive when looking for the paper source file, it must be named `paper.md`
