### 2024-04-01 e8f6a86
Fix PAPER.md caps to work with JOSS
