+++ paper.bib
@@ -65,5 +65,6 @@
 	author       = {Wan, Z. and Lo, D. and Xia, X. and Cai, L.},
 	year         = 2017,
 	booktitle    = {Proceedings of the 14th IEEE International Working Conference on Mining Software Repositories: MSR},
-	pages        = {20--21}
+	pages        = {20--21},
+	doi          = {10.1109/MSR.2017.59}
 }
