### 2024-04-02 #6565 @JIBSIL
@danielskatz - The issues with the references should have been fixed in the latest commit of paper.md
