### 2024-04-15 #6565 @JIBSIL
@aparoha - Sure! You can view it [here](https://github.com/openjournals/joss-papers/blob/joss.06565/joss.06565/10.21105.joss.06565.pdf) (or on the repo [here](https://github.com/JIBSIL/DistriFS/blob/main/paper.md))
