### 2024-04-15 #6565 @mahfuz05062
Hi @aparoha , all the reviews in JOSS are done through Github (so a little different from traditional peer-review process). I will open up a new review thread and explain the review process once we have at least two reviewers agreed to perform the review. Thanks!
