### 2024-04-15 #6625 @editorialbot
[REVIEW]: DistriFS: A Platform and User Agnostic Approach to Dataset Distribution
**Submitting author:** @JIBSIL (<a href="http://orcid.org/0009-0006-8945-0092">Julian Boesch</a>)
**Repository:** https://github.com/JIBSIL/DistriFS
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
