### 2024-04-19 JIBSIL/DistriFS#5 @suriya-ganesh
JOSS
Hi @JIBSIL ,

Thank for this project!

I'm opening this issue to track my review in your repo.
These are my part my observations and part JOSS requirements.

### JOSS Requirements

- [x] Automated unit/integration/e2e Tests for the system you've developed.
- [x] Community Guidelines , while this is a fairly straightforward project and contributing would be easy. It's nicer for new people to have these guidelines

###  Personal Observations
- The justification and failure reasons for BitTorrent as well as IPFS is very interesting and reasonable. But I'm curious, with way that DistriFS is setup right now how is it "distributed"? how would a new server coming up, discover other servers in the network? how would they share files? how does encryption work in that scenario?
- I spent some time with the code, but couldn't confirm that there were any authorization system for downloading or uploading the file. Would love to know if I'm wrong. If not, what is the logic behind having a cryptography scheme and back and forth?

JOSS issue thread: https://github.com/openjournals/joss-reviews/issues/6625
