### 2024-04-22 JIBSIL/DistriFS#5 @JIBSIL
@suriya-ganesh - To answer your questions:
- In regards to being distributed: People who wish to download from the network can operate an indexer or use somebody else's. The person who runs the indexer creates a list of indexers which their indexer receives names and hashes from, which is displayed to the client when calling the getAllFiles route (this is inefficient when many files are present in the indexer; I plan to add a different system for retrieving servers in the next release). The client can specify a filename or hash to search for, and they get a download link to the server.
- A new server (by this I assume you mean indexer/server combination) coming up would be able to connect to any popular indexer (if this system were to see popular use, I'm sure there would be popular ones or some listing) and get a recursive file listing of every file on every indexer that is known to that first indexer. Additionally, the suggestServer endpoint can be called by anyone (e.g a server listing website?) to suggest a new server to the indexer.
- Files are stored exclusively on servers. Indexers just point to them; the user discovers the file to download based on the outputs of an indexer.
- There is actually no system for uploading files sent by users. However the "passport" system allows the server owner to restrict access to certain files by public key (this is mentioned in further detail [in the server README](https://github.com/JIBSIL/DistriFS/tree/main/server#authenticating)). The server owner may then set their public key in the `AdminKeys` section or password-protect certain files through this syntax:
```yaml
PassportProtectedFiles:
  test.txt:
    - mypublickey
    - secondkey
```
(as shown in [config.go](https://github.com/JIBSIL/DistriFS/blob/main/server/modules/helper/config.go) of the server)
