### 2024-04-22 27f5959
fix: make Contributing section more clear, add support section
