### 2024-04-22 JIBSIL/DistriFS#5 @JIBSIL
I have just made some changes to the Contributing section through [this commit](https://github.com/JIBSIL/DistriFS/commit/27f5959f14bcc91936d3a19d836727a5af76345b) and [this commit](https://github.com/JIBSIL/DistriFS/commit/7681208053529a4c3d5546ebf542283b6fb6e6e0). I hope this passes the community guidelines check. I will work on rewriting my tests into native Go this week. Thanks for your review!
- JIBSIL
