### 2024-04-22 JIBSIL/DistriFS#5 @suriya-ganesh
Ah! I see, thank you for your answers.
This is great!! I didn't notice the indexer routes.
Awesome work @JIBSIL
LMK once you fix the tests, I'll approve them together and we should be good.
