### 2024-04-25 63b61d4
feat: unit testing for server and indexer
