+++ server/routes/passport/passport_test.go
@@ -1,0 +1,197 @@
+// Rather than having file-specific tests, this file contains a complete Passport test/example.
+
+package passport
+
+import (
+	"bytes"
+	"crypto/ecdsa"
+	"crypto/elliptic"
+	"crypto/rand"
+	"crypto/sha256"
+
+	// "crypto/x509"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
+	"io"
+	"math/big"
+	"net/http"
+	"testing"
+)
+
+// mostly auto-generated from Insomnia routes & responses cred: https://mholt.github.io/json-to-go/
+type AuthenticateResponse struct {
+	Data struct {
+		Key   string `json:"key"`
+		Value string `json:"value"`
+	} `json:"data"`
+	Success bool `json:"success"`
+}
+
+type GetKeyRequest struct {
+	Pubkey        string `json:"pubkey"`
+	Signedmessage string `json:"signedmessage"`
+	Messagekey    string `json:"messagekey"`
+}
+
+type GetKeyResponse struct {
+	Data    string `json:"data"`
+	Success bool   `json:"success"`
+}
+
+type VerifyKeyRequest struct {
+	Key string `json:"key"`
+}
+
+type VerifyKeyResponse struct {
+	Data    string `json:"data"`
+	Success bool   `json:"success"`
+}
+
+func GetSigningKeys(passportSigningKey string) (string, []byte) {
+	sigkey, _ := NewSigningKey()
+	sig, _ := Sign([]byte(passportSigningKey), sigkey)
+	msg := hex.EncodeToString(sig)
+	fmt.Printf("Using %s hex-signed message\n", msg)
+	// isVerified := Verify([]byte(passportSigningKey), sig, &sigkey.PublicKey)
+	// fmt.Printf("SigVerify output: %s\n", strconv.FormatBool(isVerified))
+
+	//privkey, _ := x509.MarshalECPrivateKey(sigkey)
+	// fmt.Printf("Using privkey %s\n", hex.EncodeToString(privkey))
+
+	// DECODING
+	//derBytes, _ := hex.DecodeString(hexKey)
+	//privateKey, _ := x509.ParseECPrivateKey(derBytes)
+
+	publicKey := sigkey.Public().(*ecdsa.PublicKey)
+	publicKeyBytes := elliptic.Marshal(publicKey.Curve, publicKey.X, publicKey.Y)
+	// fmt.Printf("Using pubkey %s", hex.EncodeToString(publicKeyBytes))
+
+	// DECODING
+	//fpubKeyBytes, _ := hex.DecodeString(hexKey)
+	//x, y := elliptic.Unmarshal(elliptic.P256(), pubKeyBytes)
+	//publicKey := &ecdsa.PublicKey{Curve: elliptic.P256(), X: x, Y: y}
+
+	return msg, publicKeyBytes
+}
+
+// NewSigningKey generates a random P-256 ECDSA private key.
+func NewSigningKey() (*ecdsa.PrivateKey, error) {
+	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	return key, err
+}
+
+// Sign signs arbitrary data using ECDSA.
+func Sign(data []byte, privkey *ecdsa.PrivateKey) ([]byte, error) {
+	// hash message
+	digest := sha256.Sum256(data)
+
+	// sign the hash
+	r, s, err := ecdsa.Sign(rand.Reader, privkey, digest[:])
+	if err != nil {
+		return nil, err
+	}
+
+	// encode the signature {R, S}
+	// big.Int.Bytes() will need padding in the case of leading zero bytes
+	params := privkey.Curve.Params()
+	curveOrderByteSize := params.P.BitLen() / 8
+	rBytes, sBytes := r.Bytes(), s.Bytes()
[…]
