### 2024-04-25 3e4f0df
feat: update README with testing instructions
