+++ README.md
@@ -21,5 +21,8 @@ Coming soon!
 # Contributing
 Pull requests are always open, for both bugfixes and new features! Feedback and bug reports are also open via GitHub Issues - please remember to adhere to our [Code of Conduct](CODE_OF_CONDUCT.md) in Issues. For support, please contact me through the contacts in [my bio](https://github.com/JIBSIL).
 
+# Testing
+Unit tests are present for every route on the server and indexer as of commit [63b61d4](https://github.com/JIBSIL/DistriFS/commit/63b61d4071a62edf792db49daf30bfe8cf866702). You can run the tests by using the command `go test ./...` in the `server` and `indexer` folders to test each component.
+
 # License
 DistriFS is a free-to-use academic work. It is licensed under the MIT license and is free to use in any project, commercial or noncommerical.
