### 2024-04-25 #6625 @JIBSIL
Added testing as of [commit 63b61d4](https://github.com/JIBSIL/DistriFS/commit/63b61d4071a62edf792db49daf30bfe8cf866702). Testing instructions are also added to README so users can replicate results.
