### 2024-06-13 JIBSIL/DistriFS#5 @aparoha
Thanks for adding tests and readme.
