+++ server/benchmarks/generate.sh
@@ -1,0 +1,3 @@
+fallocate -l 1G 1gb-file.bin
+fallocate -l 100M 100mb-file.bin
+fallocate -l 1M 1mb-file.bin
