+++ server/benchmarks/README.md
@@ -1,0 +1,5 @@
+## benchmarks
+
+This folder contains files to benchmark the DistriFS server, and to conduct performance comparisons.
+
+Please be advised that `start_server_benchmark.sh` will not work on every system, and is only a template for some commands that should be run to benchmark DistriFS against other comparable software.
