### 2024-06-17 #6625 @JIBSIL
@mahfuz05062 - Sure, here's the completed checklist:

Release version: [v1.0.0 @ GitHub Releases](https://github.com/JIBSIL/DistriFS/releases/tag/v1.0.0)
Archive version: [Figshare](https://figshare.com/articles/software/DistriFS/26049448), DOI 10.6084/m9.figshare.26049448
