+++ release/README.md
@@ -1,0 +1,5 @@
+## release/
+
+This folder contains all files needed to replicate binaries published to GitHub Releases.
+
+All releases are generated on Windows 11, x64.
