### 2024-06-26 #6625 @mahfuz05062
> @mahfuz05062 - It would be possible to substitute the source with a published one. However, the paper in question is the original whitepaper for the IPFS project, hence why it is so old. I found an original description of the system useful- however, if having recent published sources is an absolute requirement, I could substitute it with one of the [other academic papers on the subject](https://docs.ipfs.tech/concepts/further-reading/academic-papers/#design-and-evaluation-of-ipfs-a-storage-layer-for-the-decentralized-web).

In that case, I would suggest to add two citations for IPFS. Keep the current one and add another more recent/published citation.
