+++ paper.bib
@@ -68,3 +68,20 @@
 	pages        = {20--21},
 	doi          = {10.1109/MSR.2017.59}
 }
+@inproceedings{Trautwein:2022,
+	title        = {Design and evaluation of IPFS: a storage layer for the decentralized web},
+	author       = {Trautwein, Dennis and Raman, Aravindh and Tyson, Gareth and Castro, Ignacio and Scott, Will and Schubotz, Moritz and Gipp, Bela and Psaras, Yiannis},
+	year         = 2022,
+	booktitle    = {Proceedings of the ACM SIGCOMM 2022 Conference},
+	location     = {Amsterdam, Netherlands},
+	publisher    = {Association for Computing Machinery},
+	address      = {New York, NY, USA},
+	series       = {SIGCOMM '22},
+	pages        = {739-752},
+	doi          = {10.1145/3544216.3544232},
+	isbn         = 9781450394208,
+	url          = {https://doi.org/10.1145/3544216.3544232},
+	abstract     = {Recent years have witnessed growing consolidation of web operations. For example, the majority of web traffic now originates from a few organizations, and even micro-websites often choose to host on large pre-existing cloud infrastructures. In response to this, the "Decentralized Web" attempts to distribute ownership and operation of web services more evenly. This paper describes the design and implementation of the largest and most widely used Decentralized Web platform --- the InterPlanetary File System (IPFS) --- an open-source, content-addressable peer-to-peer network that provides distributed data storage and delivery. IPFS has millions of daily content retrievals and already underpins dozens of third-party applications. This paper evaluates the performance of IPFS by introducing a set of measurement methodologies that allow us to uncover the characteristics of peers in the IPFS network. We reveal presence in more than 2700 Autonomous Systems and 152 countries, the majority of which operate outside large central cloud providers like Amazon or Azure. We further evaluate IPFS performance, showing that both publication and retrieval delays are acceptable for a wide range of use cases. Finally, we share our datasets, experiences and lessons learned.},
+	numpages     = 14,
+	keywords     = {libp2p, interplanetary file system, decentralized web, content addressing, content addressable storage}
+}
