### 2024-06-29 #6625 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
