+++ paper.bib
@@ -50,7 +50,7 @@
 	url          = {https://doi.org/10.1145/269005.266694}
 }
 @article{Cox:2022,
-	title        = {The Go programming language and environment},
+	title        = {The {G}o programming language and environment},
 	author       = {Lu, L. and Sankaranarayana Pillai, T. and Gopalakrishnan, H. and Arpaci-Dusseau, A. C. and Cox, R. and Griesemer, R. and Pike, R. and Taylor, I. L. and Thompson, K.},
 	year         = 2022,
 	journal      = {Communications of the ACM},
