### 2024-06-30 #6625 @danielskatz
👋 @JIBSIL - as track editor, I've proofread the paper, and suggest the following changes: https://github.com/JIBSIL/DistriFS/pull/8  Please merge this, or let me know what you disagree with.

Also, I'm a bit confused by the content of [the figshare archive](https://doi.org/10.6084/m9.figshare.26049448), as I don't immediately see the source code. There's a zip file there, but I can't open it, and I see lots of other files that I don't understand being there. The intent of the archived version of the repo that we ask for is simply to archivally preserve a snapshot of [the repo](https://github.com/JIBSIL/DistriFS) at the time of publication.
