### 2024-07-02 #6625 @JIBSIL
@danielskatz - I have merged the PR for the bib/paper changes. Additionally, I was under the impression that the figshare archive was meant to contain binaries for the server and indexer as well as the source. I have corrected it in [the newest version](https://figshare.com/articles/software/DistriFS/26049448) to include just the source. It appears to show the contents of the repo properly now.
