### 2024-07-02 b86fcd4
add "cite this repository" menu when visiting the repo
