### 2024-07-02 #6625 @danielskatz
Congratulations to @JIBSIL (Julian Boesch) on your publication!!

JOSS depends on volunteers and it wouldn't be successful without you

Note: https://www.theoj.org/joss-papers/joss.06625/10.21105.joss.06625.pdf doesn't currently work for me, so I'll leave this issue open until it does. This could be a network caching issue, or a delay somewhere in the processing outside of JOSS.
