### 2024-07-02 #6625 @danielskatz
The issue is likely the maintenance mentioned in https://status.crossref.org/incidents/382r13mfjvhb
