### 2024-07-04 2e7f8af
fix: correct README.md to show the right paper links, and remove unused sections
